tweak Locations component to work with data from graphql
